+++ CHANGES.md
@@ -1,4 +1,5 @@
 ### 0.1.1-SNAPSHOT
+* feature: queries made with defquery can take arguments
 * feature: compiler encodes from java.util.Date and org.joda.time.DateTime
            formats to xsd:dateTime
 * feature: pretty-printer